+++ src/gluonts/nursery/glide/__init__.py
@@ -39,4 +39,5 @@ To help splitting the input preemptively glides offers a `partition`-method.
 __all__ = ["partition", "Map", "ParMap"]
 
 from ._partition import partition, divide_into
-from .pipeline import ParMap, Map
+from .parallel import ParMap
+from .sequential import Map
