+++ src/gluonts/nursery/glide/sequential.py
@@ -1,0 +1,25 @@
+# Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License").
+# You may not use this file except in compliance with the License.
+# A copy of the License is located at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# or in the "license" file accompanying this file. This file is distributed
+# on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+# express or implied. See the License for the specific language governing
+# permissions and limitations under the License.
+
+from toolz.itertoolz import concat
+
+
+class Map:
+    def __init__(self, fn, partitions: list, *args, **kwargs):
+        self.fn = fn
+        self.partitions = partitions
+
+    def __iter__(self):
+        yield from concat(
+            map(self.fn, partition) for partition in self.partitions
+        )
